replaces size_t with uint32_t for size in fullio routines.

+++ src/fullio.hpp
@@ -1,4 +1,4 @@
 #pragma once
 
-std::pair<bool,size_t> full_read (int fd, void *buf, size_t size);
-size_t full_write (int fd, const void *buf, size_t size);
+std::pair<bool,uint32_t> full_read (int fd, void *buf, uint32_t size);
+uint32_t full_write (int fd, const void *buf, uint32_t size);
